refactor: move message input block directly from features to widgets
