+++ client/src/shared/api/message/sendMessage.ts
@@ -8,9 +8,3 @@ export const apiSendMessage = async (message: ReadyMessage, room: number) => {
     });
     return res;
 };
-
-// END POINT HERE
-// 1. first of all move api somewhere (ask where because have no idea)
-// 4. change to slice/model/types.ts in entity
-// 5. i think i should move input panel to widget
-// 6. commit: 'refactor: move types to entities, change features folder structure'
